NEW Adds option to have relation setting show/hide toggle
Defaults to on for ManyRelationHandler and off for HasOne. When on,
setting fields aren't shown by default and there's a change relation
state button. Clicking this hides the Actions column and shows the
relation column. Then get a cancel and save relation buttons.

ManyRelationHandler has a new constructor argument for setting this.
HasOne expliciatly sets it to false on construction. Can always use
the setter method instead.